requirements.txt
supabase
python-dotenv
flet
flask
mcp
flask
flet
mcp
python-dotenv
supabase
cloudinary
rembg
requests
